+++ src/main/java/duke/DukeException.java
@@ -1,6 +1,8 @@
+package duke;
+
 public class DukeException extends Exception {
     private static final String invalidInput = "OOPS!!! I'm sorry, but I don't know what that means :(";
-    private static final String emptyDescription = "OOPS!!! The description of a task cannot be empty";
+    private static final String emptyDescription = "OOPS!!! The description of a duke.task cannot be empty";
 
     private DukeException(String message) {
         super(message);
@@ -8,7 +10,7 @@ public class DukeException extends Exception {
 
     public static DukeException invalidInput() {
         return new DukeException(invalidInput);
-    } 
+    }
 
     public static DukeException emptyDescription() {
         return new DukeException(emptyDescription);
