+++ src/main/java/duke/task/Task.java
@@ -1,4 +1,4 @@
-package task;
+package duke.task;
 
 public abstract class Task {
     protected String description;
@@ -15,7 +15,7 @@ public abstract class Task {
     }
 
     public String getStatusIcon() {
-        return (isDone ? "X" : " "); // mark done task with X
+        return (isDone ? "X" : " "); // mark done duke.task with X
     }
 
     public void markAsDone() {
